+++ tests/testthat/test_civis_future.R
@@ -1,13 +1,16 @@
 context("civis_future")
 library(civis)
 library(future)
+library(mockery)
 
 mock_r_eval <- function(fut) {
-  # anything in inst/ is actually installed in the package root.
-  # note: devtools provides a system.file that makes tests and checks pass when run interactively.
   r_remote_eval <- parse(system.file("scripts", "r_remote_eval.R", package = "civis"))
+
+  # so you can delete expressions.
+  # this deletes the commandArgs lines, which can no longer be mocked.
+  r_remote_eval[2:3] <- NULL
+
   with_mock(
-    `base::commandArgs` = function(...) list("123"),
     `civis::read_civis` = function(...) fut,
     `civis::scripts_post_containers_runs_outputs` = function(...) NULL,
     `civis::write_civis_file` = function(...) NULL,
@@ -42,13 +45,41 @@ mock_run <- function(expr) {
     `civis::scripts_post_containers_runs_outputs` = function(...) NULL,
     `civis::scripts_get_containers_runs` = function(...) list(state = "succeeded"),
     `civis::fetch_output` = function(...) mock_r_eval(fut),
-    value(run(fut))
+    `civis::fetch_logs` = function(...) list("a log"),
+    list(fut = run(fut), value = value(fut))
   )
 }
 
 test_that("run and value work", {
   out <- capture.output(res <- mock_run(quote(2 + 3)))
-  expect_equal(res, 5)
+  expect_equal(res$value, 5)
+  expect_equal(res$fut$logs, list("a log"))
+  expect_equal(res$fut$state, "succeeded")
+  # shouldn't need to be mocked
+  expect_equal(value(res$fut), res$val)
+})
+
+mock_err <- function(expr) {
+  fut <- CivisFuture(expr)
+  with_mock(
+    `civis::write_civis_file` = function(...) 123,
+    `civis::upload_runner_script` = function(...) "",
+    `civis::scripts_post_containers` = function(...) NULL,
+    `civis::scripts_post_containers_runs` = function(...) list(containerId = 1, id = 2),
+    `civis::scripts_post_containers_runs_outputs` = function(...) NULL,
+    `civis::scripts_get_containers_runs` = function(id, run_id) list(state = "failed"),
+    `civis::fetch_output` = function(...) NULL,
+    `civis::fetch_logs` = function(...) list("error_log"),
+    list(fut = run(fut), value = value(fut))
+  )
+}
+
+test_that("run and value handle errors", {
+  e <- tryCatch(mock_err(quote(2 + 3)), error = function(e) e)
+  expect_is(e, "civis_error")
+  msg <- "scripts_get_containers_runs(id = 1, run_id = 2): \nerror_log"
+  expect_equal(e$message, msg)
+  expect_equal(attributes(e)$args, list(id = 1, run_id = 2))
 })
 
 test_that("CivisFuture has the right stuff", {
@@ -69,8 +100,11 @@ test_that("CivisFuture has the right stuff", {
 })
 
 test_that("resolved", {
-  fut <- CivisFuture(quote(2 + 2))
-  expect_false(resolved(fut))
+  with_mock(
+    `civis::scripts_get_containers_runs` = function(...) list(state = "running"),
+    fut <- CivisFuture(quote(2 + 2)),
+    expect_false(resolved(fut))
+  )
 })
 
 test_that("cancel", {
@@ -81,5 +115,4 @@ test_that("cancel", {
   )
   expect_equal(fut$state, "cancelled")
 })
-