+++ tests/testthat/test_civis_ml.R
@@ -19,7 +19,6 @@ test_that("calls scripts_post_custom", {
   fake_scripts_post_custom_runs <- mock(list(id = 888))
   fake_scripts_get_custom_runs <- mock(list(state = "running"), list(state = "succeeded"))
   fake_civis_ml_fetch_existing <- mock(NULL)
-  fake_getOption <- mock(1111, cycle = TRUE)
 
   with_mock(
     `civis::get_database_id` = fake_get_database_id,
@@ -27,7 +26,6 @@ test_that("calls scripts_post_custom", {
     `civis::scripts_post_custom_runs` = fake_scripts_post_custom_runs,
     `civis::scripts_get_custom_runs` = fake_scripts_get_custom_runs,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
-    `base::getOption` = fake_getOption,
 
     tbl <- civis_table(table_name = "schema.table",
                        database_name = "a_database",
@@ -52,14 +50,14 @@ test_that("calls scripts_post_custom", {
              disk_requested = 9,
              notifications = list(successEmailSubject = "A success",
                                   successEmailAddresses = c("user@example.com")),
-             polling_interval = 5,
+             polling_interval = .01,
              validation_data = "skip",
              n_jobs = 9,
              verbose = FALSE)
   )
 
   script_args <- mock_args(fake_scripts_post_custom)[[1]]
-  expect_equal(script_args$from_template_id, 1111)
+  expect_equal(script_args$from_template_id, getOption("civis.ml_train_template_id"))
   expect_equal(script_args$name, "awesome civisml Train")
   expect_equal(script_args$notifications, list(successEmailSubject = "A success",
                                                successEmailAddresses = c("user@example.com")))
@@ -98,15 +96,11 @@ test_that("calls scripts_post_custom", {
 })
 
 test_that("calls civis_ml.data.frame for local df", {
-  fake_write_csv <- mock(NULL)
-  fake_temp_file <- mock("fake_temp_path")
   fake_write_civis_file <- mock(1234)
   fake_get_database_id <- mock(456)
   fake_create_and_run_model <- mock(NULL)
 
   with_mock(
-    `utils::write.csv` = fake_write_csv,
-    `base::tempfile` = fake_temp_file,
     `civis::write_civis_file` = fake_write_civis_file,
     `civis::get_database_id` = fake_get_database_id,
     `civis::create_and_run_model` = fake_create_and_run_model,
@@ -115,10 +109,6 @@ test_that("calls civis_ml.data.frame for local df", {
              model_type = "sparse_logistic",
              dependent_variable = "the_target_column",
              primary_key = "the_pk_column"),
-
-    expect_args(fake_write_civis_file, 1,
-                path = "fake_temp_path",
-                name = "modelpipeline_data.csv"),
 
     expect_args(fake_create_and_run_model, 1,
                 file_id = 1234,
@@ -335,14 +325,24 @@ test_that("calls scripts_post_custom", {
   fake_scripts_post_custom <- mock(list(id = 999))
   fake_scripts_post_custom_runs <- mock(list(id = 888))
   fake_scripts_get_custom_runs <- mock(list(state = "running"), list(state = "succeeded"))
+  fake_scripts_get_custom <- mock(list(state = "succeeded"), cycle = TRUE)
   fake_fetch_predict_results <- mock(NULL)
+<<<<<<< HEAD
+=======
+
+>>>>>>> upstream/master
 
   with_mock(
     `civis::get_database_id` = fake_get_database_id,
     `civis::scripts_post_custom` = fake_scripts_post_custom,
     `civis::scripts_post_custom_runs` = fake_scripts_post_custom_runs,
+    `civis::scripts_get_custom` = fake_scripts_get_custom,
     `civis::scripts_get_custom_runs` = fake_scripts_get_custom_runs,
     `civis::fetch_predict_results` = fake_fetch_predict_results,
+<<<<<<< HEAD
+=======
+
+>>>>>>> upstream/master
 
     tbl <- civis_table(table_name = "schema.table",
                        database_name = "a_database",
@@ -358,7 +358,11 @@ test_that("calls scripts_post_custom", {
             cpu_requested = 2000,
             memory_requested = 10,
             disk_requested = 15,
+<<<<<<< HEAD
             polling_interval = 5,
+=======
+            polling_interval = .01,
+>>>>>>> upstream/master
             verbose = TRUE)
   )
 
@@ -411,24 +415,15 @@ test_that("uses training primary_key by default", {
 })
 
 test_that("uploads local df and passes a file_id", {
-  fake_write_csv <- mock(NULL)
-  fake_temp_file <- mock("fake_temp_path")
   fake_write_civis_file <- mock(1234)
   fake_create_and_run_pred <- mock(NULL)
 
   with_mock(
-    `utils::write.csv` = fake_write_csv,
-    `base::tempfile` = fake_temp_file,
     `civis::write_civis_file` = fake_write_civis_file,
     `civis::create_and_run_pred` = fake_create_and_run_pred,
 
     predict(fake_model, iris, primary_key = NULL)
   )
-
-  expect_args(fake_write_csv, 1,
-              iris,
-              file = "fake_temp_path",
-              row.names = FALSE)
 
   expect_args(fake_create_and_run_pred, 1,
               train_job_id = fake_model$job$id,
@@ -579,12 +574,10 @@ test_that("passes table info", {
 context("create_and_run_model")
 
 test_that("uses the correct template_id", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
 
@@ -592,19 +585,16 @@ test_that("uses the correct template_id", {
   )
 
   run_args <- mock_args(fake_run_model)[[1]]
-  expect_equal(run_args$template_id, 999999)
+  expect_equal(run_args$template_id, getOption("civis.ml_train_template_id"))
 })
 
 test_that("converts parameters arg to JSON string", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
-
     create_and_run_model(file_id = 123, parameters = list(n_trees = 500, c = -1))
   )
 
@@ -613,12 +603,10 @@ test_that("converts parameters arg to JSON string", {
 })
 
 test_that("converts cross_validation_parameters to JSON string", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
 
@@ -633,12 +621,10 @@ test_that("converts cross_validation_parameters to JSON string", {
 })
 
 test_that("converts fit_params to JSON string", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
 
@@ -651,15 +637,12 @@ test_that("converts fit_params to JSON string", {
 })
 
 test_that("space separates excluded_columns", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
-
     create_and_run_model(file_id = 132, excluded_columns = c("c1", "c2", "c3"))
   )
 
@@ -668,12 +651,10 @@ test_that("space separates excluded_columns", {
 })
 
 test_that("space separates target_column", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
 
@@ -686,15 +667,12 @@ test_that("space separates target_column", {
 })
 
 test_that("file_id is always numeric", {
-  fake_getOption <- mock(999999, cycle = TRUE)
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
 
   with_mock(
-    `base::getOption` = fake_getOption,
     `civis::run_model` = fake_run_model,
     `civis::civis_ml_fetch_existing` = fake_civis_ml_fetch_existing,
-
     create_and_run_model(file_id = civis_file(132))
   )
 
@@ -702,6 +680,10 @@ test_that("file_id is always numeric", {
   expect_equal(run_args$arguments$CIVIS_FILE_ID, 132)
 })
 
+<<<<<<< HEAD
+=======
+
+>>>>>>> upstream/master
 test_that("exceptions with hyperband correct", {
   fake_run_model <- mock(list(job_id = 133, run_id = 244))
   fake_civis_ml_fetch_existing <- mock(NULL)
@@ -733,14 +715,19 @@ test_that("uses the correct template_id", {
     `civis::run_model` = fake_run_model,
     `civis::fetch_predict_results` = fake_fetch_predict_results,
 
+<<<<<<< HEAD
     create_and_run_pred(train_job_id = 111, train_run_id = 222, template_id = 555)
   )
 
   run_args <- mock_args(fake_run_model)[[1]]
   expect_equal(run_args$template_id, 555)
+=======
+    create_and_run_pred(train_job_id = 111, train_run_id = 222, template_id = 555),
+    run_args <- mock_args(fake_run_model)[[1]],
+    expect_equal(run_args$template_id, 555)
+  )
+>>>>>>> upstream/master
 })
-
-###############################################################################
 
 ################################################################################
 # fetch existing model
@@ -821,7 +808,8 @@ test_that("fetch_logs.civis_ml_error works", {
     `civis::scripts_post_custom_runs` = function(...) NULL,
     `civis::scripts_get_custom_runs` = function(...) list(state = "failed", id = 1, run_id = 2, error = "msg"),
     `civis::fetch_logs.civis_ml_error` = function(...) list("A log message"),
-    e <- tryCatch(civis:::run_model(1234, name = "sparse_logistic", list(), list(), verbose = TRUE),
+    e <- tryCatch(civis:::run_model(1234, name = "sparse_logistic", list(), list(),
+                                    verbose = TRUE, polling_interval = NULL),
              error = function(e) e),
     log <- fetch_logs(e)[[1]],
     expect_equal(log, "A log message"))
@@ -887,24 +875,18 @@ context("fetch_oos_scores")
 
 test_that("it checks input type", {
   fake_must_fetch_output_file <- mock(NULL)
-  fake_read_csv <- mock(NULL)
 
   with_mock(
-    `utils::read.csv` = fake_read_csv,
     `civis::must_fetch_output_file` = fake_must_fetch_output_file,
-
     expect_error(fetch_oos_scores("not a model"), "is_civis_ml(model) is not TRUE", fixed = TRUE)
   )
 })
 
 test_that("it looks for predictions.csv.gz", {
-  fake_must_fetch_output_file <- mock(NULL)
-  fake_read_csv <- mock(NULL)
+  fake_must_fetch_output_file <- mock(textConnection(c("a, b, c")))
 
   with_mock(
-    `utils::read.csv` = fake_read_csv,
     `civis::must_fetch_output_file` = fake_must_fetch_output_file,
-
     fetch_oos_scores(structure(list(), class = "civis_ml"))
   )
 
@@ -913,19 +895,15 @@ test_that("it looks for predictions.csv.gz", {
 })
 
 test_that("it calls read.csv with extra args", {
-  fake_must_fetch_output_file <- mock("a_file.csv")
-  fake_read_csv <- mock(NULL)
+  fake_must_fetch_output_file <- mock(textConnection(c("a,b,c")))
 
-  with_mock(
-    `utils::read.csv` = fake_read_csv,
+  df <- with_mock(
     `civis::must_fetch_output_file` = fake_must_fetch_output_file,
-
-    fetch_oos_scores(structure(list(), class = "civis_ml"), stringsAsFactors = FALSE)
+    fetch_oos_scores(structure(list(), class = "civis_ml"),
+                           stringsAsFactors = FALSE, header = FALSE)
   )
-
-  csv_args <- mock_args(fake_read_csv)[[1]]
-  expect_equal(csv_args[[1]], "a_file.csv")
-  expect_equal(csv_args$stringsAsFactors, FALSE)
+  ans <- data.frame(V1 = "a", V2 = "b", V3 = "c", stringsAsFactors = FALSE)
+  expect_equal(df, ans)
 })
 
 
@@ -938,18 +916,16 @@ test_that("it checks input type", {
 
 test_that("it calls read.csv with extra args, and dowload_civis with correct id", {
   fake_read_csv <- mock(NULL)
-  fake_download_civis <- mock("path.csv")
+  fake_download_civis <- mock(textConnection(c("a,b,c")))
 
-  with_mock(
-    `utils::read.csv` = fake_read_csv,
+  df <- with_mock(
     `civis::fetch_predict_results` = function(...) list(model_info = list(output_file_ids = 1)),
     `civis::download_civis` = fake_download_civis,
-    fetch_predictions(structure(list(), class = "civis_ml_prediction"), stringsAsFactors = FALSE)
+    fetch_predictions(structure(list(), class = "civis_ml_prediction"),
+                      header = FALSE, stringsAsFactors = FALSE)
   )
-
-  csv_args <- mock_args(fake_read_csv)[[1]]
-  expect_equal(csv_args[[1]], "path.csv")
-  expect_equal(csv_args$stringsAsFactors, FALSE)
+  ans <- data.frame(V1 = "a", V2 = "b", V3 = "c", stringsAsFactors = FALSE)
+  expect_equal(df, ans)
 
   dl_args <- mock_args(fake_download_civis)[[1]]
   expect_equal(dl_args[[1]], 1)