+++ tools/integration_tests/smoke_test.R
@@ -39,6 +39,17 @@ test_that("write_civis writes to redshift", {
   colnames(x) <- colnames(iris)
   expect_equivalent(x[order(x$id), ], iris)
   query_civis(paste0("DROP TABLE IF EXISTS ", tablename), database = database)
+
+  # write_civis.numeric
+  tablename <- paste0("scratch.r_smoke_test", sample(1:10000, 1))
+  fname <- tempfile(fileext = ".csv")
+  write.csv(iris, file = fname, row.names = FALSE)
+  fid <- write_civis_file(fname, "iris.csv")
+  write_civis(fid, tablename, "redshift-general", if_exists = "drop", verbose = TRUE)
+  x <- read_civis(tablename)
+  colnames(x) <- colnames(iris)
+  expect_equivalent(x[order(x$id), ], iris)
+  query_civis(paste0("DROP TABLE IF EXISTS ", tablename), database = database)
 })
 
 test_that("download_civis downloads onto disk", {
@@ -81,10 +92,20 @@ test_that("files can be uploaded", {
   x <- read_civis(fid)
   expect_equal(x, iris)
 
-  # test expires_at = NULL
-  fid2 <- write_civis_file(iris, expires_at = NULL)
-  info <- files_get(id = fid2)
+  info <- files_get(id = fid)
   expect_null(info$expiresAt)
+
+  # test large file
+  library(future)
+  plan("sequential")
+  fn <- tempfile()
+  system(paste0("dd if=/dev/zero of=", fn, " count=1 bs=",
+                civis:::MIN_MULTIPART_SIZE + 1))
+  system.time(id <- write_civis_file(fn, name = "a"))
+
+  plan("multisession")
+  # if you've got a big pipe, it can make some difference...
+  system.time(id <- write_civis_file(fn, name = "a"))
 })
 
 test_that("DBI Read-Only connections forbid writes", {